Adds ability to retry only incomplete parts & not upload files already present on S3
Adds lastModifiedDate as a comparison value to determine if upload can be resumed.

Works! Cleans up localstorage for “expired” uploads

Converts DAYS_AGO to a parameter and refactors DAYS to HOURS. Default is to NOT use the cache.

Restores missing line

Updates README.

Only sets status to 100% when status is changed to complete.

Uses a unique signature for the file key, rather than just the file name as this would cause issues if uploading two video files with the same name from different recording devices.